Add validation to abstract methods

+++ ivis-server/src/main/java/com/imcode/controllers/AbstractRestController.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.databind.JsonMappingException;
 import com.imcode.entities.interfaces.JpaEntity;
 import com.imcode.exceptions.factories.ErrorBuilder;
 import com.imcode.exceptions.wrappers.GeneralError;
+import com.imcode.misc.errors.Error;
 import com.imcode.misc.errors.ErrorFactory;
 import com.imcode.services.GenericService;
 import com.imcode.services.NamedService;
@@ -13,12 +14,18 @@ import com.imcode.utils.StaticUtls;
 import com.imcode.validators.GenericValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
+import org.springframework.core.MethodParameter;
 import org.springframework.dao.DataAccessException;
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.data.util.ReflectionUtils;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.converter.HttpMessageConversionException;
 import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.orm.jpa.JpaSystemException;
 import org.springframework.ui.Model;
+import org.springframework.util.Assert;
+import org.springframework.util.ObjectUtils;
+import org.springframework.validation.BeanPropertyBindingResult;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.WebDataBinder;
@@ -32,6 +39,7 @@ import javax.validation.Valid;
 import java.io.Serializable;
 import java.lang.reflect.InvocationTargetException;
 import java.util.HashMap;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -45,23 +53,19 @@ public abstract class AbstractRestController<T extends JpaEntity<ID>, ID extends
     @Autowired
     private SERVICE_TYPE service;
 
-    @Autowired
-    private ErrorFactory errorFactory;
-
-    @Autowired
-    private ApplicationContext ctx;
-
     // Getting entity by id
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
-    public Object get(@PathVariable("id") ID id, HttpServletResponse response, WebRequest webRequest) {
+    public Object get(@PathVariable("id") ID id, HttpServletResponse response, WebRequest webRequest) throws Exception {
         T entity = service.find(id);
+        StaticUtls.checkNullAndSetNoContent(entity, response);
         return entity;
     }
 
     //Getting list of entities
     @RequestMapping(method = RequestMethod.GET)
-    public Object getAll(WebRequest webRequest, HttpServletResponse response, Model model) {
+    public Object getAll(WebRequest webRequest, HttpServletResponse response, Model model) throws Exception {
         List<T> result = service.findAll();
+        StaticUtls.checkNullAndSetNoContent(result, response);
         return result;
     }
 
@@ -81,8 +85,15 @@ public abstract class AbstractRestController<T extends JpaEntity<ID>, ID extends
     @RequestMapping(value = "/saveall", method = RequestMethod.POST)
     public @ResponseBody Object saveAll(@RequestBody Iterable<T> entities,
                                         HttpServletResponse response,
-                                        WebRequest webRequest, @RequestParam(required = false) Boolean full) {
+                                        BindingResult bindingResult,
+                                        WebRequest webRequest, @RequestParam(required = false) Boolean full) throws MethodArgumentNotValidException {
         Iterable<T> result = service.save(entities);
+
+        Iterator<T> iterator = result.iterator();
+
+        while (iterator.hasNext()) {
+            new GenericValidator(true, "id").invoke(iterator.next(), bindingResult);
+        }
 
         if (Boolean.FALSE.equals(full)) {
             List<ID> ids = StreamSupport.stream(result.spliterator(), false).map(JpaEntity::getId).collect(Collectors.toList());
@@ -95,29 +106,40 @@ public abstract class AbstractRestController<T extends JpaEntity<ID>, ID extends
 
     // Updating entity
     @RequestMapping(value = "/{id}", method = RequestMethod.PUT)
-    public Object update(@PathVariable("id") ID id, HttpServletResponse response, @RequestBody(required = false) T entity, WebRequest webRequest) {
+    public Object update(@PathVariable("id") ID id, HttpServletResponse response, @RequestBody(required = false) T entity, BindingResult bindingResult, WebRequest webRequest) throws Exception {
         T existsEntity = getService().find(id);
 
+        new GenericValidator(true, "id").invoke(entity, bindingResult);
+
+        boolean isCopied = false;
         if (existsEntity != null) {
             try {
-                StaticUtls.nullAwareBeanCopy(existsEntity, entity);
+                isCopied =  StaticUtls.nullAwareBeanCopy(existsEntity, entity);
             } catch (IllegalAccessException e) {
                 e.printStackTrace();
             } catch (InvocationTargetException e) {
                 e.printStackTrace();
             }
-
-            service.save(existsEntity);
+            if (isCopied) {
+                return service.save(existsEntity);
+            } else {
+                StaticUtls.checkNullAndSetNoContent(null, response);
+            }
         }
 
-        return existsEntity;
-//        return service.save(entity);
+        return null;
+
     }
 
     //Deleting entity
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-    public Object delete(@PathVariable("id") ID id, HttpServletResponse response, WebRequest webRequest) {
+    public Object delete(@PathVariable("id") ID id, HttpServletResponse response, WebRequest webRequest) throws MethodArgumentNotValidException {
         T entity = service.find(id);
+        if (entity == null) {
+            BindingResult bindingResult = new BeanPropertyBindingResult(entity, "entity");
+            bindingResult.reject(null, "Try delete non exist entity");
+            throw new MethodArgumentNotValidException(null, bindingResult);
+        }
         service.delete(id);
         return entity;
     }
@@ -125,8 +147,9 @@ public abstract class AbstractRestController<T extends JpaEntity<ID>, ID extends
     @SuppressWarnings("unchecked")
 //    @RequestMapping(method = RequestMethod.GET, params = {"name"})
     public Object getByName(WebRequest webRequest, Model model,
-                         @RequestParam("name") String name,
-                         @RequestParam(value = "first", required = false) Boolean firstOnly) {
+                            HttpServletResponse response,
+                            @RequestParam("name") String name,
+                            @RequestParam(value = "first", required = false) Boolean firstOnly) {
 
         if (service instanceof NamedService) {
             NamedService<T> namedService = (NamedService<T>) service;
@@ -183,24 +206,6 @@ public abstract class AbstractRestController<T extends JpaEntity<ID>, ID extends
         return constraints;
     }
 
-//    @ExceptionHandler(MethodArgumentNotValidException.class)
-//    @ResponseStatus(value = HttpStatus.BAD_REQUEST)
-//    public GeneralError handleException(MethodArgumentNotValidException exception) {
-//        return ErrorBuilder.buildValidationError(exception.getBindingResult());
-//    }
-//
-//    @ExceptionHandler(JsonMappingException.class)
-//    @ResponseStatus(value = HttpStatus.BAD_REQUEST)
-//    public GeneralError handleException(JsonMappingException exception) {
-//        return ErrorBuilder.buildJsonMappingException(exception);
-//    }
-//
-//    @ExceptionHandler(PersistenceException.class)
-//    @ResponseStatus(value = HttpStatus.BAD_REQUEST)
-//    public GeneralError handleException(JpaSystemException exception) {
-//        return ErrorBuilder.buildDatabasePersistenceError(exception);
-//    }
-
     @ExceptionHandler(Exception.class)
     @ResponseStatus(value = HttpStatus.BAD_REQUEST)
     public GeneralError handleException(Exception exception) {
@@ -224,30 +229,4 @@ public abstract class AbstractRestController<T extends JpaEntity<ID>, ID extends
     }
 
 
-
-//    private ValidationError createValidationError(MethodArgumentNotValidException e) {
-//        return ValidationErrorBuilder.fromBindingErrors(e.getBindingResult());
-//    }
-
-
-//    @ExceptionHandler
-//    @ResponseStatus(value = HttpStatus.BAD_REQUEST)
-//    public MessageOfException messagingException(MessagingException exception) {
-//        return exception.getExceptionMeassage();
-//    }
-
-//    @ExceptionHandler
-//    @ResponseStatus(value = HttpStatus.BAD_REQUEST)
-//    public MessageOfException messagingException(Exception exception) {
-//        MessageOfException messagingException = new MessageOfException(exception);
-//        return messagingException;
-//    }
-
-//    public ErrorFactory getErrorFactory() {
-//        return errorFactory;
-//    }
-//
-//    public void setErrorFactory(ErrorFactory errorFactory) {
-//        this.errorFactory = errorFactory;
-//    }
 }
